Require std::has_unique_object_representations for SkGoodHash
Fixes one type in Graphite (SamplerDesc) that had padding bits that
theoretically could have caused hash misses. Because of the types that
use it internally to DrawPass, it was simpler to update SamplerDesc to
be directly hashable.

Adds an `SkForceDirectHash` template for key types to opt into if they
want to use `SkOpts::hash_fn` but aren't technically unique (often due
to their use of floats in the key). This was to make migrating easier
since the components that require it are in Ganesh and the PDF backend, but the intent is not to use SkForceDirectHash for new key types.

Reviewed-on: https://skia-review.googlesource.com/c/skia/+/667956

+++ modules/skparagraph/src/OneLineShaper.h
@@ -132,7 +132,7 @@ private:
         bool operator==(const FontKey& other) const;
 
         struct Hasher {
-            size_t operator()(const FontKey& key) const;
+            uint32_t operator()(const FontKey& key) const;
         };
     };
     skia_private::THashMap<FontKey, sk_sp<SkTypeface>, FontKey::Hasher> fFallbackFonts;
